add route to  productItem
