+++ src/routes/productItem.route.ts
@@ -1,0 +1,7 @@
+import express from 'express';
+import { catchError } from '../utils/catchError';
+import * as productItemController from '../controllers/productItem.controller';
+
+export const router = express.Router();
+
+router.get('/:id', catchError(productItemController.getOne));
