+++ src/server.ts
@@ -3,6 +3,7 @@
 import express from 'express';
 import cors from 'cors';
 import { router as productRouter } from './routes/product.route';
+import { router as productItem } from './routes/productItem.route';
 
 const PORT = process.env.PORT ?? 3005;
 const server = express();
@@ -10,6 +11,7 @@ const server = express();
 server.use(cors());
 
 server.use('/products', express.json(), productRouter);
+server.use('/productItem', express.json(), productItem);
 
 server.listen(PORT, () => {
   console.log(`Server is running on http://localhost:${PORT}`);
